+++ realtime-delivery/src/shadow_runtime.hpp
@@ -5,6 +5,7 @@
 #include <string>
 
 #include "health_server.hpp"
+#include "hiredis_presence_reader.hpp"
 #include "librdkafka_consumer.hpp"
 
 namespace dipole::realtime {
@@ -16,6 +17,9 @@ struct ShadowRuntimeConfig {
   int poll_timeout_ms = 250;
   int error_backoff_ms = 250;
   bool timeline_notify_shadow = false;
+  bool presence_shadow = false;
+  HiredisPresenceConfig presence;
+  std::int64_t presence_ttl_ms = 120'000;
 };
 
 ValidationError LoadShadowRuntimeConfig(ShadowRuntimeConfig* config);